Refactor MyStoreStore
