+++ internal/model/item.go
@@ -1,4 +1,4 @@
-package mysqlstorage
+package model
 
 type Item struct {
 	Id    int64
